画面遷移フローの修正
