+++ 7_Clima-iOS13/Clima/CityListView/Controller/CityListViewController.swift
@@ -13,9 +13,4 @@ class CityListViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
     }
-    @IBAction func buttonPressed(_ sender: UIButton) {
-        let storyboard = UIStoryboard(name: "Main", bundle: nil)
-        let vc = storyboard.instantiateViewController(withIdentifier: "WeatherViewController") as! WeatherViewController
-        navigationController?.pushViewController(vc, animated: true)
-    }
 }
