--- 7_Clima-iOS13/Clima/HomeView/View/HomeViewController.xib
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="32700.99.1234" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina6_12" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="22684"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="System colors in document resources" minToolsVersion="11.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="HomeViewController" customModule="Clima" customModuleProvider="target">
-            <connections>
-                <outlet property="view" destination="iN0-l3-epB" id="L6b-Eo-Ehg"/>
-            </connections>
-        </placeholder>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="iN0-l3-epB">
-            <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="5qL-be-DGu">
-                    <rect key="frame" x="129" y="400" width="117" height="35"/>
-                    <state key="normal" title="Button"/>
-                    <buttonConfiguration key="configuration" style="plain" title="お気に入り"/>
-                    <connections>
-                        <action selector="favoriteButtonPressed:" destination="-1" eventType="touchUpInside" id="hYc-dC-kof"/>
-                    </connections>
-                </button>
-            </subviews>
-            <viewLayoutGuide key="safeArea" id="vUN-kp-3ea"/>
-            <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-            <constraints>
-                <constraint firstItem="5qL-be-DGu" firstAttribute="top" secondItem="vUN-kp-3ea" secondAttribute="top" constant="341" id="USK-zF-p3d"/>
-                <constraint firstItem="vUN-kp-3ea" firstAttribute="bottom" secondItem="5qL-be-DGu" secondAttribute="bottom" constant="198" id="aOR-oo-vlE"/>
-                <constraint firstItem="5qL-be-DGu" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="129" id="xYi-ir-Cf6"/>
-                <constraint firstItem="vUN-kp-3ea" firstAttribute="trailing" secondItem="5qL-be-DGu" secondAttribute="trailing" constant="129" id="xkd-uS-88v"/>
-            </constraints>
-            <point key="canvasLocation" x="46.564885496183201" y="20.422535211267608"/>
-        </view>
-    </objects>
-    <resources>
-        <systemColor name="systemBackgroundColor">
-            <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-        </systemColor>
-    </resources>
-</document>
