+++ 7_Clima-iOS13/Clima/SceneDelegate.swift
@@ -18,7 +18,8 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         guard let scene = (scene as? UIWindowScene) else { return }
 
         let window = UIWindow(windowScene: scene)
-        let vc = HomeViewController()
+        let storyboard = UIStoryboard(name: "Main", bundle: nil)
+        let vc = storyboard.instantiateViewController(withIdentifier: "WeatherViewController") as! WeatherViewController
         let nav = UINavigationController(rootViewController: vc)
 
         window.rootViewController = nav
