+++ 7_Clima-iOS13/Clima/WeatherView/View/Base.lproj/Main.storyboard
@@ -91,8 +91,16 @@
                                         <fontDescription key="fontDescription" type="system" pointSize="30"/>
                                         <nil key="highlightedColor"/>
                                     </label>
+                                    <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="6ov-Lz-364">
+                                        <rect key="frame" x="270" y="355.5" width="104" height="34.5"/>
+                                        <state key="normal" title="Button"/>
+                                        <buttonConfiguration key="configuration" style="plain" title="お気に入り"/>
+                                        <connections>
+                                            <action selector="favoriteButtonPressed:" destination="BYZ-38-t0r" eventType="touchUpInside" id="LDj-wa-Rjd"/>
+                                        </connections>
+                                    </button>
                                     <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="0ti-bh-8Fw">
-                                        <rect key="frame" x="134" y="355.5" width="240" height="458.5"/>
+                                        <rect key="frame" x="134" y="400" width="240" height="414"/>
                                     </view>
                                 </subviews>
                                 <constraints>
